Add execution time func in model
